Rename columntypes package
